+++ shared_dataverse_information/worldmap_datatables/models.py
@@ -17,12 +17,12 @@ TABLE_JOIN_TO_RESULT_MAP = dict(tablejoin_id='pk'\
     )
 
 
+
 class TableUploadAndJoinRequest(models.Model):
 
     title = models.CharField(max_length=255, help_text='Title for New DataTable')
     table_attribute = models.CharField(max_length=255\
-                                    , help_text='DataTableattribute name to join on')
-
+                                    , help_text='DataTable attribute name to join on')
     layer_typename = models.CharField(max_length=255, help_text='Layer name')
     layer_attribute = models.CharField(max_length=255\
                                     , help_text='Layer attribute name to join on')
@@ -40,12 +40,63 @@ class TableUploadAndJoinRequest(models.Model):
         abstract = True
 
 
+class MapLatLngLayerRequest(models.Model):
+    """
+    Used for the Worldmap API request
+    """
+    title = models.CharField(max_length=255, help_text='Title for New DataTable')
+    abstract = models.TextField(default='(no abstract)')
+    lng_attribute = models.CharField(max_length=255, help_text='Longitude column name')
+    lat_attribute = models.CharField(max_length=255, help_text='Latitude column name')
+
+
+
+class MapLatLngLayerResult(models.Model):
+    """
+    Information returned after a successful mapping of a lat/lng layer
+    """
+    tablejoin_id = models.IntegerField(help_text='TableJoin pk')
+
+    tablejoin_view_name = models.CharField(max_length=255, help_text='TableJoin view_name')
+
+    # New Join Layer
+    #
+    layer_id = models.CharField(max_length=255\
+                        , help_text='Layer id')
+
+    layer_typename = models.CharField(max_length=255\
+                        , help_text='Layer typename')
+
+    layer_url = models.CharField(max_length=255\
+                        , help_text='layer.get_absolute_url()')
+
+
+    mapped_record_count = models.IntegerField(default=0, help_text='TableJoin matched_records_count')
+    unmapped_record_count = models.IntegerField(default=0, help_text='TableJoin unmatched_records_count')
+    unmapped_records_list = models.TextField(blank=True, help_text='TableJoin unmatched_records_list')
+
+    table_name = models.CharField(max_length=255, help_text='TableJoin datatable.table_name')
+    table_join_attribute = models.CharField(max_length=255\
+                        , help_text='TableJoin table_attribute.attribute')
+
+    layer_typename = models.CharField(max_length=255, help_text='TableJoin join_layer.typename')
+    layer_join_attribute = models.CharField(max_length=255\
+                        , help_text='TableJoin layer_attribute.attribute')
+
+
+    def __unicode__(self):
+        return self.join_layer_typename
+
+    class Meta:
+        abstract = True
+
+
 
 class TableJoinRequest(models.Model):
 
     table_name = models.CharField(max_length=255, help_text='DataTable name')
     table_attribute = models.CharField(max_length=255\
-                                    , help_text='DataTableattribute name to join on')
+                                    , help_text='DataTable attribute name to join on')
 
     layer_typename = models.CharField(max_length=255, help_text='Layer name')
     layer_attribute = models.CharField(max_length=255\