+++ HISTORY.rst
@@ -3,11 +3,14 @@
 Release History
 ---------------
 
-0.3.5 (2015-03-10)
+
+0.3.5 (2015-03-13)
 ++++++++++++++++++
-Added worldmap_tables to begin facilitation of geoconnect API
+Added worldmap_tables to for geoconnect API
 - Added TableJoinResult, TableJoinResultForm
 - Added TableJoinRequest, TableJoinRequestForm
+- Added MapLatLngLayerRequest, MapLatLngLayerRequestForm
+
 
 0.3.4 (2015-03-09)
 ++++++++++++++++++